+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - On quick reply, «Enter» will not send draft anymore, it must be CTRL+Enter.
 - Quick reply is now multilines.
 - Use message's excerpt in search results instead of garbled highlights
+- Move vcard file import route on apiv2 to use contact uniqueness and lookups principles
 
 ### Fixed